+++ spec/internal/app/posts/2011-05-01-full-metadata.markdown
@@ -11,3 +11,5 @@ Second paragraph of content.
 
     First line of code.
       Second line of code.
+
+<%= content_tag :p, "Paragraph created by Rails helper" %>
